// src/app/interfaces/Product.ts
// Generated by https://quicktype.io

export interface Product {
    id:        number;
    name:      string;
    description: string;
    category:  string;
    price:     number;
    size:      string;
    promotion: number;
    quantity:  number;
    image:     string;
}

export interface ProductForm {
    name:      string;
    description: string;
    category:  string;
    price:     number;
    size:      string;
    promotion: number;
    quantity:  number;
    image:     string;
}